feat: adding rag expert and retired rag expert
